Move state interfaces and shims to state.go
